requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
openai==1.10.0
packaging==24.0
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2023.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
